Various changes to UI
Added a button to go back, note that this button works as a 'ctrl-z' type of 'back', it doesn't go back to the previous slash on the url but to the previous page were the user was. Also added navigation to the layout instead of individual pages. I still need to delete the navigation component for some pages, also changed some margins and rounded borders to the navigation. Moved the button to create new protocols for the right to the left, more small changes coming