Finished styling budget generation modal
Added icons, colors and tables to the modal used for generating a new anual budget. If the protocol does not conform to the zod schema (team member must be related), the user will be asked to do this with an error message before being able to generate a new budget. If everything checks out, the user will be shown a "success" message and two tables showing him the budget items and team members, and of course, a button to finally generate the anual budget.