Feature/pq 273 sistema de mensajeria (#325)
* Added sensitive.json to .gitignore

This file contains the mongo uri for the scripts

* Added sensitive.json to .gitignore

This file contains the Mongo URI for the scripts

* Many things that are almost working

* Apparently fully finished

The queries still need work, messages pagination for example.

Todo: Implement threads (chat rooms basically).

* UI improvements

* Responsiveness

* Pagination

* UI

* Naming and removing unused variables

* Added backend folder

For now it only contains a server.js file which is the websocket server for the chat. It listens on the port 3001 of localhost, which should be replaced for an environment variable.

* Improvements for the server

* More improvements

* Removed unused import

* Added repository function and improvements in handling messages return

* Removed query client provider

There's already a provider in an upper level

* Removed defaultOptions from queryClient

* Removed package and console.log

* Fixes

* Fix in initial data population and removed local websocket server

For the dev enviroment now we're using a public wss server deployed in Glitch.com

* Adding support for notifications

* Added notifications support and fixed broadcasting messages

The websocket was broadcasting messages to every chat. If there were 2 or more chats active, meaning, the messages would appear on every one of them, this is fixed now.

Also added support for notifications.

* Removed unused import

* Notification support but not in real time

* Removed console logs and completely finished chat

* Removed unused packages

* Removed 2 more unused packages

* Updated frozen lockfile

* fiexd tsconfig

---------

Co-authored-by: bojkomatias <bojko.matias@gmail.com>