feat: Add AnualBudgetTeamMember relation to TeamMemberCategory
This commit adds the `AnualBudgetTeamMember` relation to the `TeamMemberCategory` model in the Prisma schema. The `AnualBudgetTeamMember` model now has a foreign key `categoryId` that references the `id` field of the `TeamMemberCategory` model. This allows for a one-to-many relationship between `TeamMemberCategory` and `AnualBudgetTeamMember`.

The purpose of this change is to enable the association of an annual budget with a specific team member category.

Co-authored-by: nicohorn <nicoskate000@gmail.com>