Finishied logic to edit team members
Now a protocol can be edited by the admin at all times (with a proper warning). There's a functionality where a member can be "deactivated" from a project, which sets the date when this happened. This maintains a history of who participated in the project, with the date when they started participating and when they stopped.