Feature/edit team members between convocatories (#406)
* Fixed wrong calculation in the UI

* Added comments for future improvements and fixed team members view in protocols/[id]

* Fixed budget view

* Fixed team members schema and repository. Also improved team member form list UI.
Added capability to deactivate a team member for a protocol.

* Fixed convocatory not getting automatically added

Fixed emailer and get secretaries by academic unit function

* Finishied logic to edit team members

Now a protocol can be edited by the admin at all times (with a proper warning). There's a functionality where a member can be "deactivated" from a project, which sets the date when this happened. This maintains a history of who participated in the project, with the date when they started participating and when they stopped.

* Added missing field